git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6@17354 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/colorspace-private.h
@@ -60,8 +60,9 @@ static inline void ConvertRGBToCMYK(MagickPixelPacket *pixel)
       green=QuantumScale*DecodePixelGamma(pixel->green);
       blue=QuantumScale*DecodePixelGamma(pixel->blue);
     }
-  if ((fabs(red) < MagickEpsilon) && (fabs(green) < MagickEpsilon) &&
-      (fabs(blue) < MagickEpsilon))
+  if ((fabs((double) red) < MagickEpsilon) &&
+      (fabs((double) green) < MagickEpsilon) &&
+      (fabs((double) blue) < MagickEpsilon))
     {
       pixel->index=(MagickRealType) QuantumRange;
       return;
